requirements.txt
discord.py 
python-dotenv 
better-profanity==0.6.1
dateparser
pillow
python-slugify
gradio-client
requests
google-generativeai==0.7.2
sympy